ISSUE-338 renamed services
