+++ authorization/repository/repository.go
@@ -3,7 +3,7 @@ package repository
 import (
 	"github.com/fabric8-services/fabric8-auth/account"
 	invitation "github.com/fabric8-services/fabric8-auth/authorization/invitation/repository"
-	permissionservice "github.com/fabric8-services/fabric8-auth/authorization/permission/service"
+	permissionservice "github.com/fabric8-services/fabric8-auth/authorization/permission/model"
 	resource "github.com/fabric8-services/fabric8-auth/authorization/resource/repository"
 	resourcetype "github.com/fabric8-services/fabric8-auth/authorization/resourcetype/repository"
 	role "github.com/fabric8-services/fabric8-auth/authorization/role/repository"
@@ -18,5 +18,5 @@ type Repositories interface {
 	RoleRepository() role.RoleRepository
 	IdentityRoleRepository() role.IdentityRoleRepository
 	Users() account.UserRepository
-	PermissionService() permissionservice.PermissionService
+	PermissionModelService() permissionservice.PermissionService
 }
