+++ gormapplication/application.go
@@ -7,9 +7,9 @@ import (
 	"github.com/fabric8-services/fabric8-auth/account"
 	"github.com/fabric8-services/fabric8-auth/application"
 	"github.com/fabric8-services/fabric8-auth/auth"
+	invitationservice "github.com/fabric8-services/fabric8-auth/authorization/invitation/model"
 	invitation "github.com/fabric8-services/fabric8-auth/authorization/invitation/repository"
-	invitationservice "github.com/fabric8-services/fabric8-auth/authorization/invitation/service"
-	permissionservice "github.com/fabric8-services/fabric8-auth/authorization/permission/service"
+	permissionservice "github.com/fabric8-services/fabric8-auth/authorization/permission/model"
 	resource "github.com/fabric8-services/fabric8-auth/authorization/resource/repository"
 	resourcetype "github.com/fabric8-services/fabric8-auth/authorization/resourcetype/repository"
 	scope "github.com/fabric8-services/fabric8-auth/authorization/resourcetype/scope/repository"
@@ -99,11 +99,11 @@ func (g *GormBase) InvitationRepository() invitation.InvitationRepository {
 	return invitation.NewInvitationRepository(g.db)
 }
 
-func (g *GormBase) InvitationService() invitationservice.InvitationService {
+func (g *GormBase) InvitationModelService() invitationservice.InvitationModelService {
 	return invitationservice.NewInvitationService(g)
 }
 
-func (g *GormBase) PermissionService() permissionservice.PermissionService {
+func (g *GormBase) PermissionModelService() permissionservice.PermissionService {
 	return permissionservice.NewPermissionService(g.db)
 }
